Added helpers_test.go and refactored hazelcast_controller_e2e_test.go
