+++ test/e2e/hazelcast_controller_e2e_test.go
@@ -2,17 +2,10 @@ package e2e
 
 import (
 	"bufio"
-	"bytes"
 	"context"
 	"fmt"
-	"net/http"
-	"net/url"
-	"regexp"
-	"strconv"
-	"strings"
 	"time"
 
-	hzClient "github.com/hazelcast/hazelcast-go-client"
 	hzTypes "github.com/hazelcast/hazelcast-go-client/types"
 	. "github.com/onsi/ginkgo/v2"
 	. "github.com/onsi/gomega"
@@ -20,16 +13,11 @@ import (
 	corev1 "k8s.io/api/core/v1"
 	v1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	"k8s.io/apimachinery/pkg/types"
-	"k8s.io/client-go/tools/clientcmd"
-	"k8s.io/client-go/tools/portforward"
-	"k8s.io/client-go/transport/spdy"
 	"k8s.io/utils/pointer"
 	"sigs.k8s.io/controller-runtime/pkg/client"
 
 	hazelcastcomv1alpha1 "github.com/hazelcast/hazelcast-platform-operator/api/v1alpha1"
 	n "github.com/hazelcast/hazelcast-platform-operator/controllers/naming"
-	"github.com/hazelcast/hazelcast-platform-operator/controllers/platform"
-	"github.com/hazelcast/hazelcast-platform-operator/controllers/protocol/codec"
 	codecTypes "github.com/hazelcast/hazelcast-platform-operator/controllers/protocol/types"
 	"github.com/hazelcast/hazelcast-platform-operator/test"
 	hazelcastconfig "github.com/hazelcast/hazelcast-platform-operator/test/e2e/config/hazelcast"
@@ -78,21 +66,6 @@ var _ = Describe("Hazelcast", func() {
 		assertDoesNotExist(lookupKey, &hazelcastcomv1alpha1.Hazelcast{})
 	})
 
-	create := func(hazelcast *hazelcastcomv1alpha1.Hazelcast) {
-		By("Creating Hazelcast CR", func() {
-			Expect(k8sClient.Create(context.Background(), hazelcast)).Should(Succeed())
-		})
-
-		By("Checking Hazelcast CR running", func() {
-			hz := &hazelcastcomv1alpha1.Hazelcast{}
-			Eventually(func() bool {
-				err := k8sClient.Get(context.Background(), lookupKey, hz)
-				Expect(err).ToNot(HaveOccurred())
-				return isHazelcastRunning(hz)
-			}, timeout, interval).Should(BeTrue())
-		})
-	}
-
 	createWithoutCheck := func(hazelcast *hazelcastcomv1alpha1.Hazelcast) {
 		By("Creating Hazelcast CR", func() {
 			Expect(k8sClient.Create(context.Background(), hazelcast)).Should(Succeed())
@@ -102,85 +75,48 @@ var _ = Describe("Hazelcast", func() {
 	Describe("Default Hazelcast CR", func() {
 		It("should create Hazelcast cluster", func() {
 			hazelcast := hazelcastconfig.Default(hzNamespace, ee)
-			create(hazelcast)
+			CreateHazelcastCR(hazelcast, lookupKey)
 		})
 	})
 
 	Describe("Hazelcast CR with expose externally feature", func() {
-		assertUseHazelcast := func(unisocket bool) {
-			ctx := context.Background()
-
-			By("checking Hazelcast discovery service external IP")
-			s := &corev1.Service{}
-			Eventually(func() bool {
-				err := k8sClient.Get(context.Background(), lookupKey, s)
-				Expect(err).ToNot(HaveOccurred())
-				return len(s.Status.LoadBalancer.Ingress) > 0
-			}, timeout, interval).Should(BeTrue())
-
-			addr := s.Status.LoadBalancer.Ingress[0].IP
-			if addr == "" {
-				addr = s.Status.LoadBalancer.Ingress[0].Hostname
-			}
-			Expect(addr).Should(Not(BeEmpty()))
-
-			By("connecting Hazelcast client")
-			config := hzClient.Config{}
-			config.Cluster.Network.SetAddresses(fmt.Sprintf("%s:5701", addr))
-			config.Cluster.Unisocket = unisocket
-			config.Cluster.Discovery.UsePublicIP = true
-			client, err := hzClient.StartNewClientWithConfig(ctx, config)
-			Expect(err).ToNot(HaveOccurred())
-
-			By("using Hazelcast client")
-			m, err := client.GetMap(ctx, "map")
-			Expect(err).ToNot(HaveOccurred())
-			for i := 0; i < 100; i++ {
-				_, err = m.Put(ctx, strconv.Itoa(i), strconv.Itoa(i))
-				Expect(err).ToNot(HaveOccurred())
-			}
-			err = client.Shutdown(ctx)
-			Expect(err).ToNot(HaveOccurred())
-		}
-
+		ctx := context.Background()
 		assertExternalAddressesNotEmpty := func() {
 			By("status external addresses should not be empty")
 			Eventually(func() string {
-				cr := hazelcastcomv1alpha1.Hazelcast{}
-				err := k8sClient.Get(context.Background(), lookupKey, &cr)
+				hz := &hazelcastcomv1alpha1.Hazelcast{}
+				err := k8sClient.Get(context.Background(), lookupKey, hz)
 				Expect(err).ToNot(HaveOccurred())
-				return cr.Status.ExternalAddresses
+				return hz.Status.ExternalAddresses
 			}, timeout, interval).Should(Not(BeEmpty()))
 		}
 
 		It("should create Hazelcast cluster and allow connecting with Hazelcast unisocket client", func() {
 			assertUseHazelcastUnisocket := func() {
-				assertUseHazelcast(true)
+				FillTheMapData(ctx, false, "map", 100)
 			}
-
 			hazelcast := hazelcastconfig.ExposeExternallyUnisocket(hzNamespace, ee)
-			create(hazelcast)
+			CreateHazelcastCR(hazelcast, lookupKey)
 			assertUseHazelcastUnisocket()
 			assertExternalAddressesNotEmpty()
 		})
 
 		It("should create Hazelcast cluster exposed with NodePort services and allow connecting with Hazelcast smart client", func() {
 			assertUseHazelcastSmart := func() {
-				assertUseHazelcast(false)
+				FillTheMapData(ctx, false, "map", 100)
 			}
-
 			hazelcast := hazelcastconfig.ExposeExternallySmartNodePort(hzNamespace, ee)
-			create(hazelcast)
+			CreateHazelcastCR(hazelcast, lookupKey)
 			assertUseHazelcastSmart()
 			assertExternalAddressesNotEmpty()
 		})
 
 		It("should create Hazelcast cluster exposed with LoadBalancer services and allow connecting with Hazelcast smart client", func() {
 			assertUseHazelcastSmart := func() {
-				assertUseHazelcast(false)
+				FillTheMapData(ctx, false, "map", 100)
 			}
 			hazelcast := hazelcastconfig.ExposeExternallySmartLoadBalancer(hzNamespace, ee)
-			create(hazelcast)
+			CreateHazelcastCR(hazelcast, lookupKey)
 			assertUseHazelcastSmart()
 		})
 	})
@@ -188,8 +124,7 @@ var _ = Describe("Hazelcast", func() {
 	Describe("Hazelcast cluster name", func() {
 		It("should create a Hazelcust cluster with Cluster name: development", func() {
 			hazelcast := hazelcastconfig.ClusterName(hzNamespace, ee)
-			create(hazelcast)
-
+			CreateHazelcastCR(hazelcast, lookupKey)
 			assertMemberLogs(hazelcast, "Cluster name: "+hazelcast.Spec.ClusterName)
 		})
 	})
@@ -197,17 +132,15 @@ var _ = Describe("Hazelcast", func() {
 	Context("Hazelcast member status", func() {
 
 		It("should update HZ ready members status", func() {
-			h := hazelcastconfig.Default(hzNamespace, ee)
-			create(h)
-
+			hazelcast := hazelcastconfig.Default(hzNamespace, ee)
+			CreateHazelcastCR(hazelcast, lookupKey)
 			evaluateReadyMembers(lookupKey, 3)
-
-			assertMemberLogs(h, "Members {size:3, ver:3}")
+			assertMemberLogs(hazelcast, "Members {size:3, ver:3}")
 
 			By("removing pods so that cluster gets recreated", func() {
 				err := k8sClient.DeleteAllOf(context.Background(), &corev1.Pod{}, client.InNamespace(lookupKey.Namespace), client.MatchingLabels{
 					n.ApplicationNameLabel:         n.Hazelcast,
-					n.ApplicationInstanceNameLabel: h.Name,
+					n.ApplicationInstanceNameLabel: hazelcast.Name,
 					n.ApplicationManagedByLabel:    n.OperatorName,
 				})
 				Expect(err).ToNot(HaveOccurred())
@@ -215,9 +148,9 @@ var _ = Describe("Hazelcast", func() {
 			})
 		})
 
-		It("should update HZ detailed member status", func() {
-			h := hazelcastconfig.Default(hzNamespace, ee)
-			create(h)
+		FIt("should update HZ detailed member status", func() {
+			hazelcast := hazelcastconfig.Default(hzNamespace, ee)
+			CreateHazelcastCR(hazelcast, lookupKey)
 
 			evaluateReadyMembers(lookupKey, 3)
 
@@ -262,8 +195,7 @@ var _ = Describe("Hazelcast", func() {
 				Skip("This test will only run in EE configuration")
 			}
 			hazelcast := hazelcastconfig.PersistenceEnabled(hzNamespace, "/data/hot-restart")
-			create(hazelcast)
-
+			CreateHazelcastCR(hazelcast, lookupKey)
 			assertMemberLogs(hazelcast, "Local Hot Restart procedure completed with success.")
 			assertMemberLogs(hazelcast, "Hot Restart procedure completed")
 
@@ -294,8 +226,7 @@ var _ = Describe("Hazelcast", func() {
 				Skip("This test will only run in EE configuration")
 			}
 			hazelcast := hazelcastconfig.PersistenceEnabled(hzNamespace, "/data/hot-restart")
-			create(hazelcast)
-
+			CreateHazelcastCR(hazelcast, lookupKey)
 			evaluateReadyMembers(lookupKey, 3)
 
 			By("Creating HotBackup CR")
@@ -304,13 +235,7 @@ var _ = Describe("Hazelcast", func() {
 			Expect(k8sClient.Create(context.Background(), hotBackup)).Should(Succeed())
 
 			By("Check the HotBackup creation sequence")
-			logs := test.GetPodLogs(context.Background(), types.NamespacedName{
-				Name:      hzName + "-0",
-				Namespace: hzNamespace,
-			}, &corev1.PodLogOptions{
-				Follow:    true,
-				SinceTime: &v1.Time{Time: t},
-			})
+			logs := InitLogs(t)
 			defer logs.Close()
 			scanner := bufio.NewScanner(logs)
 			test.EventuallyInLogs(scanner, timeout, logInterval).
@@ -337,8 +262,7 @@ var _ = Describe("Hazelcast", func() {
 				Skip("This test will only run in EE configuration")
 			}
 			hazelcast := hazelcastconfig.PersistenceEnabled(hzNamespace, "/data/hot-restart", false)
-			create(hazelcast)
-
+			CreateHazelcastCR(hazelcast, lookupKey)
 			evaluateReadyMembers(lookupKey, 3)
 
 			By("Creating HotBackup CR")
@@ -346,48 +270,8 @@ var _ = Describe("Hazelcast", func() {
 			hotBackup := hazelcastconfig.HotBackup(hazelcast.Name, hzNamespace)
 			Expect(k8sClient.Create(context.Background(), hotBackup)).Should(Succeed())
 
-			By("Finding Backup sequence")
-			logs := test.GetPodLogs(context.Background(), types.NamespacedName{
-				Name:      hzName + "-0",
-				Namespace: hzNamespace,
-			}, &corev1.PodLogOptions{
-				Follow:    true,
-				SinceTime: &v1.Time{Time: t},
-			})
-			defer logs.Close()
-			scanner := bufio.NewScanner(logs)
-			test.EventuallyInLogs(scanner, timeout, logInterval).
-				Should(ContainSubstring("Starting new hot backup with sequence"))
-			line := scanner.Text()
-			Expect(logs.Close()).Should(Succeed())
-
-			compRegEx := regexp.MustCompile(`Starting new hot backup with sequence (?P<seq>\d+)`)
-			match := compRegEx.FindStringSubmatch(line)
-			var seq string
-			for i, name := range compRegEx.SubexpNames() {
-				if name == "seq" && i > 0 && i <= len(match) {
-					seq = match[i]
-				}
-			}
-			if seq == "" {
-				Fail("Backup sequence not found")
-			}
-			Expect(k8sClient.Delete(context.Background(), hazelcast, client.PropagationPolicy(v1.DeletePropagationForeground))).Should(Succeed())
-
-			assertDoesNotExist(types.NamespacedName{
-				Name:      hzName + "-0",
-				Namespace: hzNamespace,
-			}, &corev1.Pod{})
-
-			By("Waiting for Hazelcast CR to be removed", func() {
-				Eventually(func() error {
-					h := &hazelcastcomv1alpha1.Hazelcast{}
-					return k8sClient.Get(context.Background(), types.NamespacedName{
-						Name:      hzName,
-						Namespace: hzNamespace,
-					}, h)
-				}, timeout, interval).ShouldNot(Succeed())
-			})
+			seq := GetBackupSequence(t)
+			RemoveHazelcastCR(hazelcast)
 
 			By("Creating new Hazelcast cluster from existing backup with 2 members")
 			baseDir := "/data/hot-restart/hot-backup/backup-" + seq
@@ -395,8 +279,7 @@ var _ = Describe("Hazelcast", func() {
 			hazelcast.Spec.ClusterSize = &[]int32{2}[0]
 			hazelcast.Spec.Persistence.DataRecoveryTimeout = 60
 			hazelcast.Spec.Persistence.AutoForceStart = true
-			create(hazelcast)
-
+			CreateHazelcastCR(hazelcast, lookupKey)
 			evaluateReadyMembers(lookupKey, 2)
 		})
 
@@ -406,8 +289,7 @@ var _ = Describe("Hazelcast", func() {
 			}
 			baseDir := "/data/hot-restart"
 			hazelcast := addNodeSelectorForName(hazelcastconfig.PersistenceEnabled(hzNamespace, baseDir, params...), getFirstWorkerNodeName())
-			create(hazelcast)
-
+			CreateHazelcastCR(hazelcast, lookupKey)
 			evaluateReadyMembers(lookupKey, 3)
 
 			By("Creating HotBackup CR")
@@ -415,48 +297,8 @@ var _ = Describe("Hazelcast", func() {
 			hotBackup := hazelcastconfig.HotBackup(hazelcast.Name, hzNamespace)
 			Expect(k8sClient.Create(context.Background(), hotBackup)).Should(Succeed())
 
-			By("Finding Backup sequence")
-			logs := test.GetPodLogs(context.Background(), types.NamespacedName{
-				Name:      hzName + "-0",
-				Namespace: hzNamespace,
-			}, &corev1.PodLogOptions{
-				Follow:    true,
-				SinceTime: &v1.Time{Time: t},
-			})
-			defer logs.Close()
-			scanner := bufio.NewScanner(logs)
-			test.EventuallyInLogs(scanner, timeout, logInterval).
-				Should(ContainSubstring("Starting new hot backup with sequence"))
-			line := scanner.Text()
-			Expect(logs.Close()).Should(Succeed())
-
-			compRegEx := regexp.MustCompile(`Starting new hot backup with sequence (?P<seq>\d+)`)
-			match := compRegEx.FindStringSubmatch(line)
-			var seq string
-			for i, name := range compRegEx.SubexpNames() {
-				if name == "seq" && i > 0 && i <= len(match) {
-					seq = match[i]
-				}
-			}
-			if seq == "" {
-				Fail("Backup sequence not found")
-			}
-			Expect(k8sClient.Delete(context.Background(), hazelcast, client.PropagationPolicy(v1.DeletePropagationForeground))).Should(Succeed())
-
-			assertDoesNotExist(types.NamespacedName{
-				Name:      hzName + "-0",
-				Namespace: hzNamespace,
-			}, &corev1.Pod{})
-
-			By("Waiting for Hazelcast CR to be removed", func() {
-				Eventually(func() error {
-					h := &hazelcastcomv1alpha1.Hazelcast{}
-					return k8sClient.Get(context.Background(), types.NamespacedName{
-						Name:      hzName,
-						Namespace: hzNamespace,
-					}, h)
-				}, timeout, interval).ShouldNot(Succeed())
-			})
+			seq := GetBackupSequence(t)
+			RemoveHazelcastCR(hazelcast)
 
 			By("Creating new Hazelcast cluster from existing backup")
 			baseDir += "/hot-backup/backup-" + seq
@@ -465,13 +307,10 @@ var _ = Describe("Hazelcast", func() {
 			Expect(k8sClient.Create(context.Background(), hazelcast)).Should(Succeed())
 			evaluateReadyMembers(lookupKey, 3)
 
-			logs = test.GetPodLogs(context.Background(), types.NamespacedName{
-				Name:      hzName + "-0",
-				Namespace: hzNamespace,
-			}, &corev1.PodLogOptions{Follow: true})
+			logs := InitLogs(t)
 			defer logs.Close()
 
-			scanner = bufio.NewScanner(logs)
+			scanner := bufio.NewScanner(logs)
 			test.EventuallyInLogs(scanner, timeout, logInterval).
 				Should(ContainSubstring("Starting hot-restart service. Base directory: " + baseDir))
 			test.EventuallyInLogs(scanner, timeout, logInterval).
@@ -486,12 +325,12 @@ var _ = Describe("Hazelcast", func() {
 			Expect(logs.Close()).Should(Succeed())
 
 			Eventually(func() *hazelcastcomv1alpha1.RestoreStatus {
-				h := &hazelcastcomv1alpha1.Hazelcast{}
+				hz := &hazelcastcomv1alpha1.Hazelcast{}
 				_ = k8sClient.Get(context.Background(), types.NamespacedName{
 					Name:      hzName,
 					Namespace: hzNamespace,
-				}, h)
-				return h.Status.Restore
+				}, hz)
+				return hz.Status.Restore
 			}, timeout, interval).Should(And(
 				Not(BeNil()),
 				WithTransform(func(h *hazelcastcomv1alpha1.RestoreStatus) hazelcastcomv1alpha1.RestoreState {
@@ -507,17 +346,17 @@ var _ = Describe("Hazelcast", func() {
 	Describe("Hazelcast Map Config", func() {
 		It("should create Map Config", func() {
 			hazelcast := hazelcastconfig.Default(hzNamespace, ee)
-			create(hazelcast)
+			CreateHazelcastCR(hazelcast, lookupKey)
 
 			m := hazelcastconfig.DefaultMap(hazelcast.Name, "map-1", hzNamespace)
 			Expect(k8sClient.Create(context.Background(), m)).Should(Succeed())
-
 			assertMapStatus(m, hazelcastcomv1alpha1.MapSuccess)
 		})
+
 		It("should create Map Config with correct default values", func() {
 			localPort := "8000"
 			hazelcast := hazelcastconfig.Default(hzNamespace, ee)
-			create(hazelcast)
+			CreateHazelcastCR(hazelcast, lookupKey)
 
 			By("port-forwarding to Hazelcast master pod")
 			stopChan, readyChan := portForwardPod(hazelcast.Name+"-0", hazelcast.Namespace, localPort+":5701")
@@ -551,7 +390,7 @@ var _ = Describe("Hazelcast", func() {
 		})
 		It("should create Map Config with Indexes", func() {
 			hazelcast := hazelcastconfig.Default(hzNamespace, ee)
-			create(hazelcast)
+			CreateHazelcastCR(hazelcast, lookupKey)
 
 			m := hazelcastconfig.DefaultMap(hazelcast.Name, "map-2", hzNamespace)
 			m.Spec.Indexes = []hazelcastcomv1alpha1.IndexConfig{
@@ -576,9 +415,10 @@ var _ = Describe("Hazelcast", func() {
 
 			// TODO: When Indexes can be decoded in the getMapConfig method, we can check if indexes are created correctly.
 		})
+
 		It("should fail when persistence of Map CR and Hazelcast CR do not match", func() {
 			hazelcast := hazelcastconfig.Default(hzNamespace, ee)
-			create(hazelcast)
+			CreateHazelcastCR(hazelcast, lookupKey)
 
 			m := hazelcastconfig.DefaultMap(hazelcast.Name, "map-3", hzNamespace)
 			m.Spec.PersistenceEnabled = true
@@ -587,10 +427,11 @@ var _ = Describe("Hazelcast", func() {
 			m = assertMapStatus(m, hazelcastcomv1alpha1.MapFailed)
 			Expect(m.Status.Message).To(Equal(fmt.Sprintf("Persistence is not enabled for the Hazelcast resource %s.", hazelcast.Name)))
 		})
+
 		It("should update the map correctly", func() {
 			localPort := "8000"
 			hazelcast := hazelcastconfig.Default(hzNamespace, ee)
-			create(hazelcast)
+			CreateHazelcastCR(hazelcast, lookupKey)
 
 			By("port-forwarding to Hazelcast master pod")
 			stopChan, readyChan := portForwardPod(hazelcast.Name+"-0", hazelcast.Namespace, localPort+":5701")
@@ -628,9 +469,10 @@ var _ = Describe("Hazelcast", func() {
 			Expect(mapConfig.MaxSizePolicy).Should(Equal(hazelcastcomv1alpha1.EncodeMaxSizePolicy[m.Spec.Eviction.MaxSizePolicy]))
 			Expect(mapConfig.EvictionPolicy).Should(Equal(hazelcastcomv1alpha1.EncodeEvictionPolicyType[m.Spec.Eviction.EvictionPolicy]))
 		})
+
 		It("should fail to update", func() {
 			hazelcast := hazelcastconfig.Default(hzNamespace, ee)
-			create(hazelcast)
+			CreateHazelcastCR(hazelcast, lookupKey)
 
 			By("creating the map config successfully")
 			m := hazelcastconfig.DefaultMap(hazelcast.Name, "map-123", hzNamespace)
@@ -652,7 +494,7 @@ var _ = Describe("Hazelcast", func() {
 			baseDir := "/data/hot-restart"
 
 			hazelcast := hazelcastconfig.PersistenceEnabled(hzNamespace, baseDir)
-			create(hazelcast)
+			CreateHazelcastCR(hazelcast, lookupKey)
 			evaluateReadyMembers(lookupKey, 3)
 
 			By("port-forwarding to Hazelcast master pod")
@@ -690,48 +532,8 @@ var _ = Describe("Hazelcast", func() {
 			hotBackup := hazelcastconfig.HotBackup(hazelcast.Name, hzNamespace)
 			Expect(k8sClient.Create(context.Background(), hotBackup)).Should(Succeed())
 
-			By("Finding Backup sequence")
-			logs := test.GetPodLogs(context.Background(), types.NamespacedName{
-				Name:      hzName + "-0",
-				Namespace: hzNamespace,
-			}, &corev1.PodLogOptions{
-				Follow:    true,
-				SinceTime: &v1.Time{Time: t},
-			})
-			defer logs.Close()
-			scanner := bufio.NewScanner(logs)
-			test.EventuallyInLogs(scanner, timeout, logInterval).
-				Should(ContainSubstring("Starting new hot backup with sequence"))
-			line := scanner.Text()
-			Expect(logs.Close()).Should(Succeed())
-
-			compRegEx := regexp.MustCompile(`Starting new hot backup with sequence (?P<seq>\d+)`)
-			match := compRegEx.FindStringSubmatch(line)
-			var seq string
-			for i, name := range compRegEx.SubexpNames() {
-				if name == "seq" && i > 0 && i <= len(match) {
-					seq = match[i]
-				}
-			}
-			if seq == "" {
-				Fail("Backup sequence not found")
-			}
-			Expect(k8sClient.Delete(context.Background(), hazelcast, client.PropagationPolicy(v1.DeletePropagationForeground))).Should(Succeed())
-
-			assertDoesNotExist(types.NamespacedName{
-				Name:      hzName + "-0",
-				Namespace: hzNamespace,
-			}, &corev1.Pod{})
-
-			By("Waiting for Hazelcast CR to be removed", func() {
-				Eventually(func() error {
-					h := &hazelcastcomv1alpha1.Hazelcast{}
-					return k8sClient.Get(context.Background(), types.NamespacedName{
-						Name:      hzName,
-						Namespace: hzNamespace,
-					}, h)
-				}, timeout, interval).ShouldNot(Succeed())
-			})
+			seq := GetBackupSequence(t)
+			RemoveHazelcastCR(hazelcast)
 
 			By("Creating new Hazelcast cluster from existing backup")
 			baseDir += "/hot-backup/backup-" + seq
@@ -740,13 +542,10 @@ var _ = Describe("Hazelcast", func() {
 			Expect(k8sClient.Create(context.Background(), hazelcast)).Should(Succeed())
 			evaluateReadyMembers(lookupKey, 3)
 
-			logs = test.GetPodLogs(context.Background(), types.NamespacedName{
-				Name:      hzName + "-0",
-				Namespace: hzNamespace,
-			}, &corev1.PodLogOptions{Follow: true})
+			logs := InitLogs(t)
 			defer logs.Close()
 
-			scanner = bufio.NewScanner(logs)
+			scanner := bufio.NewScanner(logs)
 			test.EventuallyInLogs(scanner, timeout, logInterval).
 				Should(ContainSubstring("Starting hot-restart service. Base directory: " + baseDir))
 			test.EventuallyInLogs(scanner, timeout, logInterval).
@@ -778,161 +577,3 @@ var _ = Describe("Hazelcast", func() {
 		})
 	})
 })
-
-func emptyHazelcast() *hazelcastcomv1alpha1.Hazelcast {
-	return &hazelcastcomv1alpha1.Hazelcast{
-		ObjectMeta: v1.ObjectMeta{
-			Name:      hzName,
-			Namespace: hzNamespace,
-		},
-	}
-}
-
-func isHazelcastRunning(hz *hazelcastcomv1alpha1.Hazelcast) bool {
-	return hz.Status.Phase == "Running"
-}
-
-// assertMemberLogs check that the given expected string can be found in the logs.
-// expected can be a regexp pattern.
-func assertMemberLogs(h *hazelcastcomv1alpha1.Hazelcast, expected string) {
-	logs := test.GetPodLogs(context.Background(), types.NamespacedName{
-		Name:      h.Name + "-0",
-		Namespace: h.Namespace,
-	}, &corev1.PodLogOptions{})
-	defer logs.Close()
-	scanner := bufio.NewScanner(logs)
-	for scanner.Scan() {
-		line := scanner.Text()
-		if match, _ := regexp.MatchString(expected, line); match {
-			return
-		}
-	}
-	Fail(fmt.Sprintf("Failed to find \"%s\" in member logs", expected))
-}
-
-func evaluateReadyMembers(lookupKey types.NamespacedName, membersCount int) {
-	hz := &hazelcastcomv1alpha1.Hazelcast{}
-	Eventually(func() string {
-		err := k8sClient.Get(context.Background(), lookupKey, hz)
-		Expect(err).ToNot(HaveOccurred())
-		return hz.Status.Cluster.ReadyMembers
-	}, timeout, interval).Should(Equal(fmt.Sprintf("%d/%d", membersCount, membersCount)))
-}
-
-func getFirstWorkerNodeName() string {
-	labelMatcher := client.MatchingLabels{}
-	if platform.GetPlatform().Type == platform.OpenShift {
-		labelMatcher = client.MatchingLabels{
-			"node-role.kubernetes.io/worker": "",
-		}
-	}
-	nodes := &corev1.NodeList{}
-	Expect(k8sClient.List(context.Background(), nodes, labelMatcher)).Should(Succeed())
-loop1:
-	for _, node := range nodes.Items {
-		for _, taint := range node.Spec.Taints {
-			if taint.Key == "node.kubernetes.io/unreachable" {
-				continue loop1
-			}
-		}
-		return node.ObjectMeta.Name
-	}
-	Fail("Could not find a reachable working node.")
-	return ""
-}
-
-func addNodeSelectorForName(hz *hazelcastcomv1alpha1.Hazelcast, n string) *hazelcastcomv1alpha1.Hazelcast {
-	// If hostPath is not enabled, do nothing
-	if hz.Spec.Scheduling == nil {
-		return hz
-	}
-	// If NodeSelector is set with dummy name, put the real node name
-	if hz.Spec.Scheduling.NodeSelector != nil {
-		hz.Spec.Scheduling.NodeSelector = map[string]string{"kubernetes.io/hostname": n}
-	}
-	return hz
-}
-
-func createHazelcastClient(ctx context.Context, h *hazelcastcomv1alpha1.Hazelcast, localPort string) *hzClient.Client {
-	config := hzClient.Config{}
-	cc := &config.Cluster
-	cc.Name = h.Spec.ClusterName
-	cc.Network.SetAddresses("localhost:" + localPort)
-	client, err := hzClient.StartNewClientWithConfig(ctx, config)
-	Expect(err).To(BeNil())
-	return client
-
-}
-
-func getMapConfig(ctx context.Context, client *hzClient.Client, mapName string) codecTypes.MapConfig {
-	ci := hzClient.NewClientInternal(client)
-	req := codec.EncodeMCGetMapConfigRequest(mapName)
-	resp, err := ci.InvokeOnRandomTarget(ctx, req, nil)
-	Expect(err).To(BeNil())
-	return codec.DecodeMCGetMapConfigResponse(resp)
-}
-
-func portForwardPod(sName, sNamespace, port string) (chan struct{}, chan struct{}) {
-	defer GinkgoRecover()
-	stopChan, readyChan := make(chan struct{}, 1), make(chan struct{}, 1)
-
-	rules := clientcmd.NewDefaultClientConfigLoadingRules()
-	kubeConfig := clientcmd.NewNonInteractiveDeferredLoadingClientConfig(rules, &clientcmd.ConfigOverrides{})
-	config, err := kubeConfig.ClientConfig()
-	Expect(err).To(BeNil())
-
-	roundTripper, upgrader, err := spdy.RoundTripperFor(config)
-	Expect(err).To(BeNil())
-
-	path := fmt.Sprintf("/api/v1/namespaces/%s/pods/%s/portforward", sNamespace, sName)
-	hostIP := strings.TrimPrefix(config.Host, "https://")
-	serverURL := url.URL{Scheme: "https", Path: path, Host: hostIP}
-	dialer := spdy.NewDialer(upgrader, &http.Client{Transport: roundTripper}, http.MethodPost, &serverURL)
-
-	out, errOut := new(bytes.Buffer), new(bytes.Buffer)
-
-	forwarder, err := portforward.New(dialer, []string{port}, stopChan, readyChan, out, errOut)
-	Expect(err).To(BeNil())
-
-	go func() {
-		if err := forwarder.ForwardPorts(); err != nil { // Locks until stopChan is closed.
-			GinkgoWriter.Println(err.Error())
-			Expect(err).To(BeNil())
-		}
-	}()
-
-	return stopChan, readyChan
-
-}
-
-func waitForReadyChannel(readyChan chan struct{}, dur time.Duration) error {
-	timer := time.NewTimer(dur)
-	for {
-		select {
-		case <-readyChan:
-			return nil
-		case <-timer.C:
-			return fmt.Errorf("Timeout waiting for readyChannel")
-		}
-	}
-}
-func closeChannel(closeChan chan struct{}) {
-	closeChan <- struct{}{}
-}
-
-func assertMapStatus(m *hazelcastcomv1alpha1.Map, st hazelcastcomv1alpha1.MapConfigState) *hazelcastcomv1alpha1.Map {
-	checkMap := &hazelcastcomv1alpha1.Map{}
-	By("Waiting for Map CR status", func() {
-		Eventually(func() hazelcastcomv1alpha1.MapConfigState {
-			err := k8sClient.Get(context.Background(), types.NamespacedName{
-				Name:      m.Name,
-				Namespace: m.Namespace,
-			}, checkMap)
-			if err != nil {
-				return ""
-			}
-			return checkMap.Status.State
-		}, timeout, interval).Should(Equal(st))
-	})
-	return checkMap
-}
